scripts/calc_engine/requirements.txt
numpy>=1.24
pandas>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pyyaml>=6.0
